Create new Discovery class
